requirements.txt
flask~=0.11
flask-restful==0.3.5
gevent==1.1.1
routes==2.3.1
webargs==1.3.4
uWSGI == 2.0.13.1
mongoengine
requests
pyyaml
